Move GithubClient to github.py.
For consistency with all the other services, move all github code into
one module.
